# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/lider": {
            "post": {
                "description": "Cadastra um novo líder no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Líderes"
                ],
                "summary": "Cria um novo líder",
                "parameters": [
                    {
                        "description": "Dados do líder",
                        "name": "lider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Lider"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lider"
                        }
                    },
                    "400": {
                        "description": "Falha ao cadastrar líder",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/lider/{cpf}": {
            "get": {
                "description": "Busca um líder pelo CPF informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Líderes"
                ],
                "summary": "Exibe um líder por CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do líder",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lider"
                        }
                    },
                    "404": {
                        "description": "Líder não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um líder do sistema pelo CPF informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Líderes"
                ],
                "summary": "Deleta um líder por CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do líder",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Líder deletado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Líder não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Falha ao deletar o líder",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Atualiza os dados de um líder existente pelo CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Líderes"
                ],
                "summary": "Atualiza um líder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do líder",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do líder",
                        "name": "lider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Lider"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lider"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos para atualização",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Líder não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Falha ao atualizar o líder",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Lider": {
            "description": "Representa um líder com informações básicas.",
            "type": "object",
            "properties": {
                "cpf": {
                    "description": "CPF do líder",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Data de criação",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Data de exclusão (opcional)",
                    "type": "string"
                },
                "email": {
                    "description": "Email do líder",
                    "type": "string"
                },
                "nome": {
                    "description": "Nome do líder",
                    "type": "string"
                },
                "senha": {
                    "description": "Senha do líder",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Data de atualização",
                    "type": "string"
                }
            }
        }
    }
}